Switch to API Token-based search
Fixes #3

Prior to this PR, the YouTube Hubot package would return a `devicesupport` link indicating that the search functionality was broken because it was using an outdated API. This moves the search to V3 of the YouTube API and restores functionality. It will require the addition of `YOUTUBE_API_KEY` environment variable.